Fail early if time is behind on agent
Let's say time is out of sync on the node to be puppetized and
it's 11:23 while on the puppetmaster it's in sync and it's 12:23.

The certificate request generated by puppet-agent will be from
11:23 but the CA on the puppetmaster will generate a certificate that is valid
from 12:23.

This PR avoids creating a certificate request on the puppet agent because
fixing this means fixing time, removing on the agent and revoking the
certificate on the puppetmaster.

If there's longer than 60 seconds between gathering facts of the node to be
puppetized and the task in this role that gets the time from the
puppetmaster, then the new variable {{ puppetize_time_difference }} may increased.